% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{corels}
\alias{corels}
\title{Corels interace}
\usage{
corels()
}
\value{
A constant bool for now
}
\description{
R Interface to 'Certifiably Optimal RulE ListS (Corels)'
}
\details{
Nothing here yet.
}
